+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/firAnonymousFunctionExpression.kt
@@ -41,4 +41,5 @@ public fun FirSession.nothingFirAnonymousFunctionExpression(
                 this.returnTypeRef = buildResolvedTypeRef {
                     this.type = builtinTypes.nothingType.coneType
                 }
-            })
+            }
+    )