+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/context/receivers/IrGeneration.kt
@@ -13,7 +13,7 @@ public fun interface IrGeneration {
         object : IrGenerationExtension {
             override fun generate(
                 moduleFragment: IrModuleFragment,
-                pluginContext: IrPluginContext
+                pluginContext: IrPluginContext,
             ) {
                 with(moduleFragment) { with(pluginContext) { generate() } }
             }