+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/declarationIrBuilder.kt
@@ -5,10 +5,8 @@ import org.jetbrains.kotlin.backend.common.lower.DeclarationIrBuilder
 import org.jetbrains.kotlin.ir.declarations.IrSymbolOwner
 import org.jetbrains.kotlin.ir.symbols.IrSymbol
 
-public fun IrPluginContext.declarationIrBuilder(
-    symbol: IrSymbol,
-): DeclarationIrBuilder = DeclarationIrBuilder(this, symbol)
+public fun IrPluginContext.declarationIrBuilder(symbol: IrSymbol): DeclarationIrBuilder =
+    DeclarationIrBuilder(this, symbol)
 
-public fun IrPluginContext.declarationIrBuilder(
-    element: IrSymbolOwner,
-): DeclarationIrBuilder = DeclarationIrBuilder(this, element.symbol)
+public fun IrPluginContext.declarationIrBuilder(element: IrSymbolOwner): DeclarationIrBuilder =
+    DeclarationIrBuilder(this, element.symbol)