+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irValueParameter.kt
@@ -10,4 +10,5 @@ public fun IrValueParameter.toIrGetValue(): IrGetValue =
         startOffset = UNDEFINED_OFFSET,
         endOffset = UNDEFINED_OFFSET,
         type = symbol.owner.type,
-        symbol = symbol)
+        symbol = symbol,
+    )