+++ kotlin-compiler-test-extensions/main/kotlin/com/javiersc/kotlin/compiler/test/services/AdditionalFilesDirectives.kt
@@ -12,5 +12,6 @@ public object AdditionalFilesDirectives : SimpleDirectivesContainer() {
                     Adds common context annotations
                     See file ./test-data/additional-files/SomeFile.kt
                 """
-                    .trimIndent())
+                    .trimIndent()
+        )
 }