+++ CHANGELOG.md
@@ -14,8 +14,8 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:hubdle-version-catalog -> 0.3.5`
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.7.6`
+- `com.javiersc.hubdle:hubdle-version-catalog -> 0.3.6`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.7.7`
 
 ## [0.2.0+2.0.10] - 2024-08-16